+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## v1.1.0 / 2021-02-25
+
+- [FEATURE] Add node resource limits watcher
+- [ENHANCEMENT] Add summary and details to incident reference crd
+
 ## v1.0.0 / 2021-02-23
 
 - [FEATURE] Add pod termination and waiting watcher